Camera controls & BG image

+++ headers/Graphics/Bindable/IndexBuffer.h
@@ -3,15 +3,15 @@
 
 class IndexBuffer : public Bindable {
 public:
-	IndexBuffer(Graphics& gfx, const std::string& tag, const u16* pData, usize count, bool updatable = false, u32 maxCount = 0);
+	IndexBuffer(Graphics& gfx, const std::string& tag, const u32* pData, usize count, bool updatable = false, u32 maxCount = 0);
 
-	void Update(Graphics& gfx, const u16* data, usize count);
+	void Update(Graphics& gfx, const u32* data, usize count);
 
 	void Bind(Graphics& gfx) noexcept override;
 	u32 GetCount() const noexcept;
 
-	static std::shared_ptr<IndexBuffer> Resolve(Graphics& gfx, const std::string& tag, const u16* pData, usize count, bool updatable = false, u32 maxCount = 0);
-	static std::string GenerateUID(const std::string& tag, const u16* pData, usize count, bool updatable = false, u32 maxCount = 0) {
+	static std::shared_ptr<IndexBuffer> Resolve(Graphics& gfx, const std::string& tag, const u32* pData, usize count, bool updatable = false, u32 maxCount = 0);
+	static std::string GenerateUID(const std::string& tag, const u32* pData, usize count, bool updatable = false, u32 maxCount = 0) {
 		UNREFERENCED_PARAMETER(pData);
 		UNREFERENCED_PARAMETER(count);
 		return GenerateUID_(tag, updatable, maxCount);
